# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Vessl API Documentation",
        "title": "Vessl API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/teams/{teamId}/audit-logs": {
            "get": {
                "description": "ListAuditLogs endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "ListAuditLogs endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "teamId",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/teams/{teamId}/ssh-keys": {
            "get": {
                "description": "ListSSHKeys endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "ListSSHKeys endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "teamId",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "post": {
                "description": "CreateSSHKey endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "CreateSSHKey endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "teamId",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.CreateSSHKeyRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/teams/{teamId}/trusted-domains": {
            "get": {
                "description": "ListTrustedDomains endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "ListTrustedDomains endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "teamId",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "post": {
                "description": "CreateTrustedDomain endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "CreateTrustedDomain endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "teamId",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.CreateTrustedDomainRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/workspaces": {
            "post": {
                "description": "Create endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workspaces"
                ],
                "summary": "Create endpoint",
                "parameters": [
                    {
                        "description": "Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.CreateWorkspaceRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/workspaces/{id}": {
            "get": {
                "description": "Get endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Get endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "teamId",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "description": "Update endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workspaces"
                ],
                "summary": "Update endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vessl_dev_vessl_internal_models.Workspace"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "description": "Delete endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workspaces"
                ],
                "summary": "Delete endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "internal_handlers.CreateSSHKeyRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "publicKey": {
                    "type": "string"
                }
            }
        },
...
        "internal_handlers.CreateTrustedDomainRequest": {
            "type": "object",
            "properties": {
                "domain": {
                    "type": "string"
                }
            }
        },
        "internal_handlers.CreateWorkspaceRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "vessl_dev_vessl_internal_models.Workspace": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "ownerId": {
                    "type": "string"
                },
                "preferredRegion": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}